Clean up

+++ src/main/scala/com/tmnow/CommentService.scala
@@ -8,8 +8,8 @@ object CommentService {
 
   def insert(row: CommentRow): IO[Int] = ???
   def getAll: IO[List[CommentRow]] = ???
-  def getById(id: Int): IO[Option[CommentRow]] = ???
-  def update(id: Int, name: String): IO[Int] = ???
-  def delete(id: Int): IO[Int] = ???
+  def getById(id: Long): IO[Option[CommentRow]] = ???
+  def update(id: Long, name: String): IO[Int] = ???
+  def delete(id: Long): IO[Int] = ???
 
 }
